backend/requirements.txt
# Karachuonyo Website Backend Dependencies

# Core Flask framework
Flask==2.3.3
Werkzeug==2.3.7

# CORS support for frontend integration
Flask-CORS==4.0.0

# Email functionality
Flask-Mail==0.9.1

# Database (SQLite is built-in to Python)
# For production, consider PostgreSQL:
# psycopg2-binary==2.9.7
# Flask-SQLAlchemy==3.0.5

# Security and validation
WTForms==3.0.1
Flask-WTF==1.1.1

# Environment variables
python-dotenv==1.0.0

# HTTP client for external API calls
requests==2.31.0

# Date and time utilities
python-dateutil==2.8.2

# JSON Web Tokens for authentication (future use)
PyJWT==2.8.0

# Password hashing (already in Werkzeug, but explicit)
bcrypt==4.0.1

# Production WSGI server
gunicorn==21.2.0

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0

# Logging and monitoring
flask-limiter==3.5.0

# Data validation
marshmallow==3.20.1

# Environment-specific configurations

# M-Pesa Daraja API integration
requests-oauthlib==1.3.1
cryptography==41.0.7
base64==0.1.0
Flask-Migrate==4.0.5  # For database migrations in production